+++ packages/svelte-query-persist-client/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @tanstack/svelte-query-persist-client
 
+## 6.0.20
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.19
+  - @tanstack/svelte-query@6.0.18
+
+## 6.0.19
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.18
+  - @tanstack/svelte-query@6.0.17
+
+## 6.0.18
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.17
+  - @tanstack/svelte-query@6.0.16
+
 ## 6.0.17
 
 ### Patch Changes